Stablized Builder / added custom write fn
Builder provides a simple interface for constructing data from parts,
and is generic enough that if the underlying data colation method needs
to change writing will be unaffected

Added write_formatted_output which is the first bash at using Builder to
write out data

Moved RegexOn to Field, expanding its use to be a more general enum for
cheaply refering to output fields.

Added Delimiter enum which attempts to minimize allocations for the
delimiter, without sacrificing usability.

Ported an iterator adapter from aureate for detecting first and last
items in an iter.

Various improvements to logging