Added env and file configs
Handrolled the env config trawler, as clap's restrictions around flag
envs was troublesome. File is expected to be .toml, but it will be
reasonably easy to added my deserialize targets in the future. Currently
only looks in /etc but that will be improved in the coming days.

Although the components are hooked into the logging, the logger does not
initialize until later leading to silent failures which will also be
fixed in the near future